move is_papermill back and check for atty instead
